move bsp APIs to JavaModule
