+++ scalalib/src/mill/scalalib/JavaModule.scala
@@ -1393,6 +1393,64 @@ trait JavaModule
       )
     }
   }
+
+  def bspBuildTargetSources: Task[(Seq[os.Path], Seq[os.Path])] = Task.Anon {
+    Tuple2(sources().map(_.path), generatedSources().map(_.path))
+  }
+
+  def sanitizeUri(uri: String): String =
+    if (uri.endsWith("/")) sanitizeUri(uri.substring(0, uri.length - 1)) else uri
+
+  def sanitizeUri(uri: os.Path): String = sanitizeUri(uri.toNIO.toUri.toString)
+
+  def sanitizeUri(uri: PathRef): String = sanitizeUri(uri.path)
+
+  def bspBuildTargetInverseSources[T](id: T, searched: String): Task[Seq[T]] = Task.Anon {
+    val src = allSourceFiles()
+    val found = src.map(sanitizeUri).contains(searched)
+    if (found) Seq(id) else Seq()
+  }
+
+  def bspBuildTargetDependencySources(includeSources: Boolean) = Task.Anon {
+    val repos = allRepositories()
+    val buildSources = if (!includeSources) Nil
+    else mill.scalalib.Lib
+      .resolveMillBuildDeps(repos, None, useSources = true)
+      .map(sanitizeUri(_))
+
+    (
+      millResolver().classpath(
+        Seq(
+          coursierDependency.withConfiguration(coursier.core.Configuration.provided),
+          coursierDependency
+        ),
+        sources = true
+      ),
+      unmanagedClasspath(),
+      buildSources
+    )
+  }
+
+  def bspBuildTargetDependencyModules = Task.Anon {
+    (
+      // full list of dependencies, including transitive ones
+      millResolver()
+        .resolution(
+          Seq(
+            coursierDependency.withConfiguration(coursier.core.Configuration.provided),
+            coursierDependency
+          )
+        )
+        .orderedDependencies,
+      unmanagedClasspath()
+    )
+  }
+
+  def bspBuildTargetScalaMainClasses = Task.Anon((allLocalMainClasses(), forkArgs(), forkEnv()))
+
+  def bspRun(args: Seq[String]): Command[Unit] = Task.Command {
+    run(Task.Anon(Args(args)))()
+  }
 }
 
 object JavaModule {
